backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.3.1
argon2-cffi==23.1.0
PyJWT==2.8.0
python-multipart==0.0.9
slowapi==0.1.9
redis==5.0.6
httpx==0.27.0
